app-engine/front/requirements.txt
Flask==1.0.2
google-cloud-datastore==1.7.3
google-cloud-secret-manager==0.2.0
pytz==2019.1
sentry-sdk==0.14.2
